requirements.txt
chromadb
sentence-transformers
streamlit
Pillow